Move coredns related functions to dns.go.
